Fixed persistent webcam usage
Fixed error where webcam would stay on after a page change, as the MediaStreamTracks are never stopped. Note: it cannot be set to do so after clicking "Stop Recording" as the MediaStream is still needed if the user wants to record again, and I could not find a way to get window.onbeforeunload to work with Turbolinks so I settled with my current implementation.